Switched back to a default alb_sph of 1 when zbrent solver fails
The algorithm first tries to solve the implicit equation describing the atmosphere-surface system. Sometimes, the solver does not converge. At some point I was assigning isnow = -<band nr for which solver failed> to identify these cases. But I apparently made a typo because it ended up into alb_sph = - <band nr for which solver failed>.

Now it's back to a default value of 1 when solver fails (compatible with Alex's fortran code and to the code sent to BC).

@RasmusBahbah I think this solves the discrepancy between EDC and BC outputs.

I have saved some values of pixels for which the solver was failing and will try to find out why it fails.